pyproject.toml
[tool.poetry]
name = "becksie-server"
version = "0.1.0"
description = ""
authors = ["Simon Palmer <palmer.simond@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
locust = "^2.32.4"
faster-whisper = "^1.1.0"
runpod = "1.7.0"
ctranslate2 = "4.4.0"
numba = "~0.56.0"
nemo-toolkit = {version = ">=2.dev", extras = ["asr"]}
beam-client = "^0.2.124"
pybind11 = "^2.13.6"
texterrors = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
